+++ Packages/io.chainsafe.web3-unity.web3auth/Runtime/WalletGUI/Prefabs/Web3AuthWalletGUI.prefab
@@ -4969,8 +4969,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   txManager: {fileID: 7234845548737740600}
   guiManager: {fileID: 804371200852392299}
-  acceptRequestButton: {fileID: 9165754778767219097}
-  rejectRequestButton: {fileID: 8654605469475103381}
   autoConfirmTxLabel: {fileID: 2012475564352559022}
   walletIconContainer: {fileID: 2788637392479046008}
   primaryBackgroundObjects:
@@ -5082,6 +5080,8 @@ MonoBehaviour:
   incomingTxNotification: {fileID: 4706101805464217323}
   txHistoryPlaceHolder: {fileID: 861225078783303682}
   txHistoryDisplay: {fileID: 8139811105131153425}
+  acceptRequestButton: {fileID: 9165754778767219097}
+  rejectRequestButton: {fileID: 8654605469475103381}
   walletGui: {fileID: 7656622686525558909}
   guiManager: {fileID: 804371200852392299}
 --- !u!114 &804371200852392299
